6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 /**
- * main - A program that prints the size of various types on the computer types
- * REturn 0 (Sccess)
+ * main -A program that prints the size of various types on the computer types
+ * return 0 (Success)
  */
 int main(void)
 {
@@ -10,11 +10,10 @@ int main(void)
 	long int c;
 	long long int d;
 	float f;
-
-printf("Size of a char: %lu byte(s)\n* (unsigned long)sizeof(a));
-printf("Siz of an int: %lu byte(s)\n*,(unsigned long)sizeof(b));
-printf("Size of a long int: %lu bytes(s)\n*, (unsigned long)sizeof(c));
-printf("Siz of a long long int: %lu bytes(s)\n*, (unsigned long)sizeof(d));
-printf("Size of a float: %lu bytes(s)\n*, (unsigned long)sideof(f));
+printf("Size of a char : % lu byte(s)\n *, sizeof(a));
+printf("Siz of an int : % lu byte(s)\n *, sizeof(b));
+printf("Size of a long int : % lu bytes(s)\n *, sizeof(c));
+printf("Siz of a long long int : % lu bytes(s)\n *, sizeof(d));
+printf("Size of a float : % lu bytes(s)\n *, sideof(f));
 return (0);
 }
